Preparar el proyecto para presentación pública
Corrige los fallos que aparecían al clonar o usar el sistema, añade la
cobertura de tests que los fija y reescribe el README como pieza de
portafolio con capturas del sistema real.

Correcciones:
- registro de modelos: guardaba la ruta absoluta de la máquina donde se
  entrenó, así que al clonar o mover el proyecto `load_active_artifact()`
  fallaba con FileNotFoundError. Ahora guarda solo el nombre y resuelve
  contra models_dir (con fallback para registros antiguos).
- pipeline diario: `process_pending_filings` importaba torch de forma
  eager y tumbaba el run entero cuando el extra opcional 'sentiment' no
  estaba instalado. Ahora degrada igual que la macro cuando FRED falla.
- universo: Wikipedia movió la lista del NASDAQ-100 fuera del artículo
  "Nasdaq-100", así que `refresh-universe` fallaba siempre. Se apunta a la
  página correcta y se toma "ICB Industry" como sector de relleno, con lo
  que los 12 tickers que no estaban en el S&P 500 dejan de quedarse sin
  sector.
- API: recargar /oportunidades, /portafolio o /salud devolvía un 404 JSON
  porque StaticFiles no hace fallback de SPA. Los 404 de /api siguen
  siendo JSON.
- frontend: `new Date('2026-08-24')` es medianoche UTC, así que todas las
  fechas de calendario se mostraban un día antes en cualquier huso al
  oeste de Greenwich. Mismo problema en la fecha por defecto al registrar
  una posición.
- portafolio: el aviso de concentración se dispara siempre con menos de 10
  posiciones porque se mide sobre el capital ya invertido; ahora solo se
  evalúa cuando el límite es alcanzable.
- npm: 5 vulnerabilidades (4 altas) resueltas con `npm audit fix`.

Calidad:
- Ruff configurado y en CI; todas las reglas seleccionadas pasan.
- Vitest en el frontend, que no tenía tests, y su paso en CI.
- 60 tests de backend y 7 de frontend, incluidos los de regresión de todo
  lo anterior.
- CI en Node 22 (20 está en EOL), `engines` declarado y Dependabot.
- .gitattributes para fijar finales de línea LF.
- recharts fuera del bundle inicial: 800 kB -> 460 kB (246 -> 146 kB gzip).

Documentación:
- README nuevo: problema, capturas del sistema real, diagrama de la
  arquitectura, rigor metodológico, métricas por fold del modelo activo,
  estructura, puesta en marcha, tabla de invariantes cubiertas por tests,
  limitaciones y resumen en inglés.
- frontend/README.md era la plantilla por defecto de Vite; ahora documenta
  scripts, estructura y convenciones reales.

+++ backend/screener/pipeline.py
@@ -1,12 +1,11 @@
 """Orquestación de los flujos del sistema (invocados desde el CLI y la API)."""
 import traceback
+from collections.abc import Iterator
 from contextlib import contextmanager
-from typing import Iterator
 
 import pandas as pd
 
-from screener.config import ensure_dirs, settings
-
+from screener.config import ensure_dirs
 
 # Plan de fases por tipo de run: (etiqueta, peso ≈ duración relativa). Permite
 # convertir "fase actual + sub-progreso" en un porcentaje GLOBAL 0-100 sobre el
@@ -64,7 +63,7 @@ class RunProgress:
         if total and total > 0 and current is not None:
             frac = max(0.0, min(1.0, current / total))
         acc = sum(self.weights[: self.idx]) + self.weights[self.idx] * frac
-        overall = int(round(100 * acc / self.total_w))
+        overall = round(100 * acc / self.total_w)
         # nunca retrocede y se reserva el 100% para el cierre del run
         overall = max(self.last_overall, min(overall, 99))
         self.last_overall = overall
